Made a wrapper class for the state enum so it can be passed about
Made a wrapper class for the state enum  so it can be passed about
because of the way java passes values through parameters. To change
the state do this:

Pass the state object (which is an instance of the StateClass class)
into your class then just do state.state = State.YOURCHOOSENSTATE

Messy but it works.